Group integration tests
in order to make it possible to include/exclude certain tests based on
the featureset of the certain region we can make use of ansible-test
utility and add group into aliases file for every integration test. This
way it is possible to invoke them as `ansible-test integration
autoscaling/` or `ansible-test integration --exclude autoscaling/`.
This also gives possibility to separate other things (dns
private/public) since depending on the region we have them both or only
one of them